Applied patch from HARMONY-4270 [drlvm] Fixes for compilation of drlvm by intel compiler on windows/ia32
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@549824 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/optimizer/abcd/classic_abcd.cpp
@@ -120,7 +120,7 @@ IOpndProxy::IOpndProxy(int32 c, uint32 id) :
     IOpnd(0, false /* is_phi */, true /* is_constant */),
     _opnd(NULL)
 {
-    setID(min_const_opnd + id);
+    setID((uint32)min_const_opnd + id);
     setConstant(c);
 }
